editoast: fix clone cli
CLI was making a shallow duplication instead of duplicating the data too
